corrected the assignment error

+++ components/event/org.wso2.carbon.event.core/src/main/java/org/wso2/carbon/event/core/sharedmemory/SharedMemoryDeliveryManagerFactory.java
@@ -43,7 +43,7 @@ public class SharedMemoryDeliveryManagerFactory implements DeliveryManagerFactor
         int minSpareThreads = Integer.parseInt(JavaUtil.getValue(config, EB_DM_MIN_SPARE_THREADS));
         int maxThreads = Integer.parseInt(JavaUtil.getValue(config, EB_DM_MAX_THREADS));
         int maxQueuedRequests = Integer.parseInt(JavaUtil.getValue(config, EB_DM_MAX_QUEUED_REQUESTS));
-        long keepAliveTime = Integer.parseInt(JavaUtil.getValue(config, EB_DM_KEEP_ALIVE_TIME));
+        int keepAliveTime = Integer.parseInt(JavaUtil.getValue(config, EB_DM_KEEP_ALIVE_TIME));
 
         String topicStoragePath = JavaUtil.getValue(config, EB_DM_TOPIC_STORAGE_PATH);
         BlockingQueue<Runnable> queue = new ArrayBlockingQueue<Runnable>(maxQueuedRequests);
